+++ README.md
@@ -1,10 +1,1 @@
 # Chatroom Apps
-
-## Chatroom with Go Channel
-[chatroom-channel](./chatroom-channel)
-
-## Chatrooom with Redis Pub/Sub
-// TODO
-
-## Chatroom with RabbitMQ
-// TODO
