--- chatroom-channel/.gitignore
@@ -1,2 +1,0 @@
-chatroom-channel
-client-app
