+++ chatroom/model/message.go
